Refactor imports for improved readability
