+++ src/app.module.ts
@@ -1,5 +1,5 @@
 import { MiddlewareConsumer, Module, NestModule } from '@nestjs/common';
-import { ConfigModule, ConfigService } from '@nestjs/config';
+import { ConfigModule } from '@nestjs/config';
 import { APP_INTERCEPTOR } from '@nestjs/core';
 
 import { CommonModule } from './common/common.module';
@@ -13,8 +13,6 @@ import multiTenantConfig from './multi-tenant/multi-tenant.config';
 import { RequestResponseInterceptor } from './logger/request-response.interceptor';
 import { ArticlesModule } from './articles/articles.module';
 import { AuthModule } from './auth/auth.module';
-import { JwtModule } from '@nestjs/jwt';
-import { EnvironmentVars } from './common/common.types';
 
 @Module({
   imports: [
