+++ src/auth/auth.module.ts
@@ -2,11 +2,12 @@ import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
 
-import { AuthService } from './auth.service';
-import { JwtStrategy } from './strategies/jwt.strategy';
-import { AuthController } from './auth.controller';
-
 import { EnvironmentVars } from 'src/common/common.types';
+
+import { AuthController } from './auth.controller';
+import { AuthService } from './auth.service';
+
+import { JwtStrategy } from './strategies/jwt.strategy';
 
 @Module({
   providers: [AuthService, JwtStrategy],
@@ -17,6 +18,7 @@ import { EnvironmentVars } from 'src/common/common.types';
       useFactory: async (configService: ConfigService) => ({
         secret: configService.get<string>(EnvironmentVars.JWT_SECRET),
         signOptions: { expiresIn: '86400s' },
+        // signOptions: { expiresIn: '20s' },
       }),
     }),
   ],
