+++ src/auth/dto/index.ts
@@ -1,3 +1,4 @@
 export * from './login.dto';
 export * from './permissions';
 export * from './profile';
+export * from './signup.dto';
